Fix platform selection on truffleruby
By using the `force_ruby_platform` dependency attribute and making it
have the right value according to the dependency name on truffleruby.

We actually had a spec for this case, but was not catching the problem
because Truffleruby monkeypatches RubyGems with

```
Gem.platforms = [Gem::Platform::RUBY]
```

and our previous implementation of the `simulate_platform` helper was
undoing that.

By fixing the helper, our existing specs start covering the issue
properly.