Don't unlock dependencies of gemspecs when gemspecs change
If the case of a gemspec changing for example its version number, there
was a comment about needing to unlock the gem. However, we were going
directly to process the next spec without unlock it, nor adding it to
the list of `converged` specs (which is what allows its dependencies to
be processed when finding the set of locked specs that should be passed
to the resolver).

This commit adds the spec to the `@unlock[:gems]` array explicitly, and
removes the `next` so that the logic to add it to the `converged` set
also applies here.