Make sure selected candidates can actually be installed
When running in legacy lockfile mode (only RUBY platform in lockfile),
it could happen that a specific candidate for the current platform is
selected that's not actually installable for the current Ruby due to not
matching Ruby version requirements. Check this case before hand so that
we can use the generic version instead.