Fix platform specific gems removed from the lockfile
If we have a lockfile missing platform a specific gem for the current
platform, and that gem happens to be already installed locally, then
Bundler will initially try to use "local mode" and use the existing
lockfile exclusively. But then when materializing it will realize that
it's missing platform gems and try to re-resolve to fill them in.
However, since we initially used "local mode", this resolution will only
consider local gems and will result in gems for other platforms being
lost.

To fix this, we need to save the resolution mode in the definition early
and make sure we set up sources using that mode before re-resolving.