Fix `bundler/inline` not resolving properly if gems not preinstalled
In inline mode, we resolve twice. First, we do a local resolution, and
if that fails (either because local resolution failed, or because some
resolved spec is not installed), we try again remotely.

Due to a recent refactoring, we were eagerly passing source requirements
for resolution to a resolution context that would be reused for both of
the above resolutions. However, the source requirements are different in
the above two modes, because local resolution uses a global aggregated
source, instead of explicit sources for each name.

The solution is to change local resolution to also use explicit source
requirements.