Preserve relative path sources in standalone setup
What was the end-user or developer problem that led to this PR?
--

Fixes https://github.com/rubygems/rubygems/issues/5317

When specifying gems with the `:path` option, `bundler --standalone`
generates a `bundler/setup.rb` with absolute paths for those gems.

This makes good sense when the `:path` is absolute (the behavior was
explicitly added in https://github.com/rubygems/rubygems/pull/4792 to
handle absolute paths).

But when the `:path` option is relative, it is not always desirable to
have absolute paths in `bundler/setup.rb`. For example, I work on an
application that runs `bundler --standalone` in CI to build a deploy
artifact, but because the root directory in CI does not match the root
in production we are unable to use the generated bundler/setup.rb
directly.

What is your fix for the problem, implemented in this PR?
--

This commit reverts back to bundler-setup-relative paths in if the
`:path` option is relative. It preserves the current behavior of using
absolute paths only when the `:path` option is absolute.