TarReader::Entry fix for rewind when reading data.tar.gz in a gem
Zlib::GzipReader doesn't support #pos= so you can't rewind a
TarReader::Entry.

Gem::Package#open_tar_gz is used in many places to read the data.tar.gz
contained within a .gem file. This normally works if you never rewind or
try to use TarReader#seek. TarReader::Entry wraps the gzio when reading
from data.tar.gz, but attempting to rewind or seek within the gzip
fails. The only solution supported by Zlib::GzipReader is to rewind and
read back to the orig_pos.

This would not be encountered during normal extract of a gem, but does
happen when someone uses the methods on Gem::Package to try to access
individual files within the gem, or to rewind an entry that's already
been read.