Sort by hash when expiring sources
Currently, git sources have a `#to_s` method that may leave side
effects, since it actually "materialize" the source (it checks it out and
figures out HEAD revision, to include it when printing the source).

This caused some surprising breakage when I was moving code around, and
it's not necessary as long as the sorting is consistent.