Handle realworld specs lost when dropping Ruby 2.7
Unfortunately I couldn't find a non-realworld scenario when introducing
these specs to prove fixes, and they were specific to ruby 2.7
versions.

After dropping Ruby 2.7 support, coverage was lost.

For one of them, I'm assuming the coverage loss because the Gemfile was
very particular to Ruby 2.7, and I wouldn't know how to do better now,
the fix happened a long time ago. We could try patching the current ruby
to fake it's Ruby 2.7, but not sure if it's worth the effort.

For the other one, I moved it to a spec that runs for all rubies and just
makes sure that the Gemfile can be resolved in "reasonable" time.