Introduce Bundler::UI::JsonShell for --json
-- Removes unreachable code from print_gem --
This method is only called (now) from inside 'if options[:parseable]',
so we can drop the check inside it - the other branches are unreachable.
And that means that groups_text is also droppable.

-- Rename print_gem/print_gems to be more specific --
These used to be the only gem-printing methods, but now they are purely
used for the 'parseable' printing path - they should be named to make
that obvious; we ended up with print_gem(s)_as_porcelain.

-- Extract table from Cli::Outdated to UI::Shell
Make it a _slightly_ general-purpose/reusable facility.
* It takes a headers Hash and Array of row Hashes
* keys in the headers that are missing from rows are left blank
* keys in rows that aren't in the headers are left out
* values are to-string'ed for display; enumerable values are joined with
  commas first
* Doesn't do any terminal size-detection stuff, just assumes everything
  will fit.
* Add table method to UI::Silent

-- Add the UI::JsonShell
* acts just like UI::Shell (by extending it), but routes all output
  through stderr instead of stdout.
* updates 'table' to produce json output on stdout.
* supports a ':pretty => true' option, which uses pretty_generate

-- Simplified some of the code that handles sorting/filtering by groups
it's an array of strings now, instead of a joined string, which meant
rewroking a bit of the logic that interacted with it

-- Introduced `--json` option for commands listed as supporting it --
When supplied and supported, use the UI::JsonShell instead of UI::Shell
as the ui class for the cli. Since only the 'table' method actually
produces json, "supporting json output" might be not be trivial for some
tools, but it might be for others - let me know if there are tools where
it would make sense to add and I'll dig in again!