Allow `extconf.rb` to cancel building the extension
Ref: https://bugs.ruby-lang.org/issues/20152

There are various gems that ship with a native extension as a way to
speedup part of the gem, but also ship with a pure Ruby version of
these methods as a fallback. So they only want to compile the extension
if the platform supports it, and if not, just fallback Ruby implementation.

Right now users rely on one of two hacks to do this. Either they create
an empty Makefile, but then still depend on make being available and it
feels very hacky, or they publish platform specific packages without any
extension in them.

Examples include `bootnsap`, `erb`, `hiredis-client`, `ddtrace`, `prism`.

This changes attempt to make this use case a first class citizen by
checking if the `extconf.rb` did generate a `gem.build_skipped` file.
If it did, `make` isn't invoked at all.