Fix relative path bug in ext_conf_builder
https://github.com/rubygems/rubygems/pull/6137 introduced a bug in
`get_relative_path` that will sometimes cause gem extensions to get
installed to the wrong location.

We found this while trying to install the google-protobuf gem, which has
its extension in `ext/google/protobuf_c`. The problem is that `zip` will
only zip up to the length of `path`, but in the case of google-protobuf
`base` has more components than `path`.

So with this base and path:

```
/home/.rubies/ruby-master/lib/ruby/gems/3.4.0+0/extensions/x86_64-darwin-23/3.4.0+0-static/google-protobuf-3.25.2
/home/.rubies/ruby-master/lib/ruby/gems/3.4.0+0/gems/google-protobuf-3.25.2/ext/google/protobuf_c
```

That will `zip` to:

```
[["home", ..., ["google-protobuf-3.25.2", "google"]]
```

rather than:

```
[["home", ..., ["google-protobuf-3.25.2", "google"], [nil, "protobuf_c"]
```

As a result, the extension gets built into the wrong directory (with
four "../" instead of five). The gem appears to install just fine, but
then fails to load because the file isn't where it should be.

We think there was also a theoretical (although unlikely) bug with
`next result if path_component == base_component`, since components that
aren't part of the prefix could also match. If a gem had an
extension dir like "ext/google-protobuf-3.25.2", for example, we'd end
up stripping that away incorrectly.

This commit reworks `get_relative_path` to avoid these bugs. First, we
pad `path_parts` with `nil` so we can safely `zip`. Then, instead of
dropping any components that match we use `drop_while` to only drop
matching components that come at the beginning.

Co-authored-by: Gabriel Amaral <gabriel-amaral@github.com>
Co-authored-by: David Jensenius <djensenius@github.com>