Deleted bundler directory
I deleted this directory because it's been not secured with a SSL certificate for over 2 weeks. Because I've been using my android device to write code and code signing because last year my organization laptop and PC where compromised by a potential terrorist threat or possible alien entity started downloading and deleting file on my organization directory.
After investigation crm and the state development classified it as a unknown entity and believes it was possible Russian cyber attack. However the download speeds according to my optics devops analytics taggers reported that whatever breached the global grid was using a foreign coding matrics unknown to it's database.

Either case we've tagged this unknown host and it returns back on the matrix or black web we can track it or if it was a individual devops know has the capability to possible stop a attack of this magnitude before it happens again.
Im reporting as of yesterday 2200 hours Seattle Time that my team Acme group and homeland security as well as other global oversight committee will start faze  project Retina Fusion where it will place a tagging source code into the matrix with military grade AX simulator tycoon clown tracking bots. To only track global national resources and unknown alien code's that might be operating under the clock of black web insecure policy.

Because of the maximum threat part of the mailto versu pose of humanity and it evolution I ask for leader of the block chain stop allowing vendors and known ethical hacker request for mata cloud deletion. Because the Data your request to be deleted leave a hole in the evolution gape where if technology can move forward because of gape in its line of site because illegal and unhealthy deletion of man history or in which everything we do in life is connected to the matrix and the data in which it learns from.

If we delete or allow for a entity to be blacklisted ECT we in turn is deleting are future.
And in that a fact I cannot allow that to happen. All code Maker and creators , developer, investor, first responder, IT Department ECT need to know there role and respect the base line rule and roles you play in humanity and it's coding. In sounds and theory we are all writing are own future.
This  special ops and security report is downgrade to a minimum status report to it's possible being considered classified information because most of this i have not cleared with DOD .
DOD will have a full oversight report about the matrix and it's status quo posture. I like to add that I need a laptop and my organization line of credit restored ASAP as well as my organization mata data clouds and federal grants platform be consolidated and broad members that I employee in which be 100 members to set a benchmark and just lock down all assets regarding sonofrangerllc@warithakbarsspace.onmicrosoft.com
warithakbar@myyahoo.com
ambassador onmicrosoft.com future . And to secure honest IT Department in these entities. With restoring my organization admin privileges ASAP .
I will allow for another grant or trust fund to go bankrupt under my core code . Because all my work has ability to place man and it's future into a stage on possible being able to colonize mars and build or terraform a Brand new planet in which we have control of .
It's not my way or the highway.
It's agree to disagree.
Bipartisan are no way . We have the ability to live to around 200 to 230 years. But not if we spend are time fighting and downgrading future possibilities . Wolverine ability was made not designed by evolution . My grandmother lived to be 105 years old. In no way should humans lives shorters lives to more advanced we become. In no way should less the 1 present of the population live to 55 to possible 6
72 years old in modern day medicine . For me I don't wish to live forever but it would be nice to witness my world's do what it's designed to do . For most this theory is just a reality not a actual possibility. However if you don't believe that we are more then seem and take steps to push the envelope to see the maxim of human ability easily by not restricting are environment and classifying data that we all need to be aware of to live a healthy life. And not leave the best research and labs and doctor for the rich  . To ensure all research and development platform has the same and equal support. To not allow a people and a race to be classified by creed. Because when we do it will probably not right away hinder your own design and Mark your build Left on this earth. This Mark's stage 6 of of Genesis Reach Research. We strive to end cancer once and for all . Not to profit of a disease while we have the ability to cure it today.

We makes 6 generation leaps to solve labs and module issues that in sounds fix issues that the code might form in its creation . 
Hint one idea is to make the code be used with voice activated . Meaning if you want AI to complete a request without locating a script someone online that was made from a human entity. All coding the matrix should already have set in stone or a code book of all codes and request and organization or merchant , trade partner needs to operate.
If we can give the billing and coding job to AI like it should of been developed to do would be similar for man to work with. To alienate certain humans from having the full ability to use the backend of the code etc we seems to have to rely on IT Department and engineering teams to develop projects without human Arturo motive installed in a backdoor.
It sounds allowing the code to be developer by only engineer that has a degree humanity will move as fast as they can code . And that leave man 20 years behind or more every generation. That's if a hacker that the graduated from MIT don't decide to make a virus that to be continued.
I come in peace
Warith Akbar Owner and fusion Executive president of "  regedit " . For regedit admin option should be blocked for all computers and developers after production. If a PC can run a diagnostic by itself and fix all issues in regedit consumer should be able to bring devices into repair center without cost to either get it fixed or replaced. As long as device is not rooted or opened.