Explicily lock turbo_tests to 2.2.0 in Gemfile
Ruby-core CI is having issues with turbo_tests 2.2.1. This version adds
json as a dependency and the `bundle install` command ruby-core uses is
not able to compile it for some reason.

I was not able to reproduce the issue locally, so this needs more
investigation.

We're already locked to 2.2.0, but ruby-core does not use a
`Gemfile.lock` so we need to lock explicitly in the Gemfile for now.